Moved functions into the diyepw package
