+++ lib/diyepw/diyepw/__init__.py
@@ -1,1 +1,5 @@
 from .typical_meteorological_year import TypicalMeteorologicalYear
+from .functions import clean_noaa_df, \
+    convert_to_station_pressure, \
+    handle_missing_values, \
+    split_list_into_contiguous_segments
